requirements.txt
google-genai
python-dotenv
numpy